+++ lib/src/storage/v3/package_dependency_storage_v3.g.dart
@@ -1,0 +1,23 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: non_constant_identifier_names
+
+part of 'package_dependency_storage_v3.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_PackageDependencyStorageV3 _$$_PackageDependencyStorageV3FromJson(
+        Map<String, dynamic> json) =>
+    _$_PackageDependencyStorageV3(
+      packageName: json['packageName'] as String,
+      packageVersion: json['packageVersion'] as String,
+    );
+
+Map<String, dynamic> _$$_PackageDependencyStorageV3ToJson(
+        _$_PackageDependencyStorageV3 instance) =>
+    <String, dynamic>{
+      'packageName': instance.packageName,
+      'packageVersion': instance.packageVersion,
+    };