+++ test/integration_tests/helper/integration_test_helper.dart
@@ -1,4 +1,5 @@
 import 'package:dart_apitool/api_tool.dart';
+import 'package:test/test.dart';
 
 class PackageApiRetriever {
   final String packageName;
@@ -16,4 +17,17 @@ class PackageApiRetriever {
     print('Analyzing $packageName $packageVersion');
     return await analyzer.analyze();
   }
+}
+
+expectPackageDependency(
+    PackageApi packageApi, String packageName, String packageVersion) {
+  final packageDependency =
+      packageApi.packageDependencies.cast<PackageDependency?>().firstWhere(
+            (element) => element != null && element.packageName == packageName,
+            orElse: () => null,
+          );
+  expect(packageDependency, isNotNull,
+      reason: 'Package dependency $packageName not found');
+  expect(packageDependency!.packageVersion, packageVersion,
+      reason: '$packageName doesn\'t have version $packageVersion');
 }