+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - changing required interfaces (even adding things) is now considered a breaking change
   interfaces are considered 'required' if they are abstract and used in a public method or field.
   This indicates that this interfaces is intended to be implemented by the user and therefore adding things also is breaking the API.
+- detects changes in package dependencies
 
 ## Version 0.7.0
 - *BREAKING*: renamed "Class" model elements to "Interface" to reflect the abstraction it represents